refactor:
파일 위치 변경 및 테스트 코드 파일 위치 리펙토링
