+++ core/src/test/java/stadium/StadiumEntityTest.java
@@ -1,4 +1,4 @@
-package com.goti.domain.entity.stadium;
+package stadium;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.test.context.ActiveProfiles;
 
+import com.goti.domain.entity.stadium.StadiumEntity;
 import com.goti.exception.FieldValidationException;
 
 import lombok.extern.slf4j.Slf4j;
