+++ core/src/test/java/user/MemberEntityTest.java
@@ -1,4 +1,4 @@
-package com.goti.domain.entity.user;
+package user;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -11,6 +11,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.springframework.test.context.ActiveProfiles;
 
 import com.goti.constants.Gender;
+import com.goti.domain.entity.user.MemberEntity;
 import com.goti.exception.FieldValidationException;
 
 import lombok.extern.slf4j.Slf4j;
