+++ integration/src/main/java/com/goti/config/api/interceptor/ExternalLoggingInterceptor.java
@@ -1,29 +1,25 @@
 package com.goti.config.api.interceptor;
 
-import org.springframework.http.HttpHeaders;
+import java.io.IOException;
+
 import org.springframework.http.HttpRequest;
 import org.springframework.http.client.ClientHttpRequestExecution;
 import org.springframework.http.client.ClientHttpRequestInterceptor;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.stereotype.Component;
 
-import java.io.IOException;
-import java.util.UUID;
-
 @Component
-public class ExternalTraceInterceptor implements ClientHttpRequestInterceptor {
-	private static final String TRACE_HEADER = "X-Trace-Id";
+public class ExternalLoggingInterceptor implements ClientHttpRequestInterceptor {
 
 	@Override
 	public ClientHttpResponse intercept(
 		HttpRequest request, byte[] body, ClientHttpRequestExecution execution
 	) throws IOException {
 
-		HttpHeaders headers = request.getHeaders();
-		if (!headers.containsKey(TRACE_HEADER)) {
-			headers.add(TRACE_HEADER, UUID.randomUUID().toString());
-		}
+		long start = System.nanoTime();
+		ClientHttpResponse response = execution.execute(request, body);
+		long tookMs = (System.nanoTime() - start) / 1_000_000;
 
-		return execution.execute(request, body);
+		return response;
 	}
 }
