+++ integration/src/main/java/com/goti/config/properties/RestClientProperties.java
