+++ integration/src/test/java/com/goti/infra/api/base/BaseRestClientTest.java
@@ -1,11 +1,9 @@
 package com.goti.infra.api.base;
 
 import com.goti.config.api.RestClientConfig;
-
 import com.goti.config.api.interceptor.ExternalLoggingInterceptor;
-
 import com.goti.config.api.interceptor.ExternalTraceInterceptor;
-
+import com.goti.config.properties.RestClientProperties;
 import com.goti.constants.messages.ErrorCode;
 import com.goti.exception.CustomException;
 
@@ -13,16 +11,21 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.client.RestClient;
 
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @Import({
 	RestClientConfig.class,
@@ -31,11 +34,18 @@ import static org.junit.jupiter.api.Assertions.*;
 })
 @Slf4j
 @ActiveProfiles("test")
-@SpringBootTest
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = { RestClientConfig.class })
+@EnableConfigurationProperties(RestClientProperties.class)
+@TestPropertySource(properties = {
+	"external.api.connect-timeout=5",
+	"external.api.read-timeout=10"
+})
 public class BaseRestClientTest {
 
 	@Autowired
 	RestClient restClient;
+
 
 	static final String BASE_URL = "https://jsonplaceholder.typicode.com";
