+++ user/src/test/java/com/goti/GotiUserApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class GotiApplicationTests {
+class GotiUserApplicationTests {
 
 	@Test
 	void contextLoads() {
