+++ user/src/test/resources/application-test.yml
@@ -1,0 +1,7 @@
+spring:
+  config:
+    import:
+      - optional:file:.env.test[.properties]
+  jpa:
+    hibernate:
+      ddl-auto: create-drop
